requirements.txt
tensorflow==1.7.0
numpy==1.14.3
pandas==0.22.0
opencv-python==3.4.0.12
scikit-learn==0.19.1
Pillow==5.1.0